+++ src/Pure.Primitives.Materialized/Pure.Primitives.Materialized.csproj
@@ -15,4 +15,8 @@
     <PackageProjectUrl>https://github.com/kudima03/Pure.Primitives.Materialized</PackageProjectUrl>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Pure.Primitives.Abstractions" Version="4.0.2" />
+  </ItemGroup>
+
 </Project>
