--- src/Pure.Primitives/Date/MaterializedDate.cs
@@ -1,26 +1,0 @@
-﻿using Pure.Primitives.Abstractions.Date;
-using System;
-
-namespace Pure.Primitives.Date;
-
-public sealed record MaterializedDate
-{
-    private readonly IDate _value;
-
-    public MaterializedDate(IDate value)
-    {
-        _value = value;
-    }
-
-    public DateOnly Value => new DateOnly(_value.Year.NumberValue, _value.Month.NumberValue, _value.Day.NumberValue);
-
-    public override int GetHashCode()
-    {
-        throw new NotSupportedException();
-    }
-
-    public override string ToString()
-    {
-        throw new NotSupportedException();
-    }
-}
