--- src/Pure.Primitives/Time/MaterializedTime.cs
@@ -1,31 +1,0 @@
-﻿using Pure.Primitives.Abstractions.Time;
-using System;
-
-namespace Pure.Primitives.Time;
-
-public sealed record MaterializedTime
-{
-    private readonly ITime _value;
-
-    public MaterializedTime(ITime value)
-    {
-        _value = value;
-    }
-
-    public TimeOnly Value => new TimeOnly(
-        _value.Hour.NumberValue,
-        _value.Minute.NumberValue,
-        _value.Second.NumberValue,
-        _value.Millisecond.NumberValue,
-        _value.Microsecond.NumberValue);
-
-    public override int GetHashCode()
-    {
-        throw new NotSupportedException();
-    }
-
-    public override string ToString()
-    {
-        throw new NotSupportedException();
-    }
-}
